+++ src/MancalaDiffEvaluator.java
@@ -6,9 +6,9 @@ public class MancalaDiffEvaluator implements BoardEvaluator {
 
   }
 
-  public int evaluateBoard(Board board, Side side) {
-    int kalahsDiff = board.getSeedsInStore(side) - board.getSeedsInStore(side.opposite());
-    int seedsDiff = board.getNumbersOfSeedsInPits(side) - board.getNumbersOfSeedsInPits(side.opposite());
+  public double evaluateBoard(Board board, Side side) {
+    double kalahsDiff = board.getSeedsInStore(side) - board.getSeedsInStore(side.opposite());
+    double seedsDiff = board.getNumbersOfSeedsInPits(side) - board.getNumbersOfSeedsInPits(side.opposite());
 
     return kalahsDiff*2 + seedsDiff;
   }
