+++ src/OptimizeResult.java
@@ -2,14 +2,14 @@ package MKAgent;
 
 public class OptimizeResult {
   public int hole;
-  public int score;
+  public double score;
 
-  public OptimizeResult(int hole, int score) {
+  public OptimizeResult(int hole, double score) {
     this.hole = hole;
     this.score = score;
   }
 
-  public OptimizeResult(int score) {
+  public OptimizeResult(double score) {
     this.score = score;
     this.hole = 0;
   }
